+++ climt/__init__.py
@@ -7,12 +7,14 @@ from ._core import (
     get_interface_values,
     climt_quantity_descriptions,
     bolton_q_sat, bolton_dqsat_dT,
-    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic)
+    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic,
+    ClimtSpectralDynamicalCore)
 
 from ._components import (
     Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation, HeldSuarez,
     GridScaleCondensation, BergerSolarInsolation, SimplePhysics, RRTMGLongwave, RRTMGShortwave,
-    EmanuelConvection, SlabSurface, IceSheet)
+    EmanuelConvection, SlabSurface, IceSheet, GfsDynamicalCore,
+    DcmipInitialConditions)
 
 sympl.add_direction_names(z=['interface_levels', 'mid_levels'])
 
@@ -21,8 +23,9 @@ __all__ = (
     get_interface_values,
     bolton_q_sat, bolton_dqsat_dT, climt_quantity_descriptions,
     ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic,
+    ClimtSpectralDynamicalCore,
     Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation, HeldSuarez,
     GridScaleCondensation, BergerSolarInsolation, SimplePhysics, RRTMGLongwave, RRTMGShortwave,
-    EmanuelConvection, SlabSurface, IceSheet)
+    EmanuelConvection, SlabSurface, IceSheet, GfsDynamicalCore, DcmipInitialConditions)
 
 __version__ = '0.9'