+++ climt/_components/gfs/__init__.py
@@ -1,0 +1,3 @@
+from .component import GfsDynamicalCore
+
+__all__ = (GfsDynamicalCore)